--- deployment/azure/.gitignore
@@ -1,1 +1,0 @@
-azure-deployment-configuration.sh
