--- deployment/vagrant-development/Vagrantfile
@@ -1,28 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-
-require_relative '../vagrant-common/checks'
-require_relative '../vagrant-common/core'
-
-provider = get_provider
-check_provider_needs(provider)
-
-Vagrant.configure("2") do |config|
-
-  config.vm.provision :file do |file|
-      file.source = "../vagrant-common/custom_motd.sh"
-      file.destination = "~/.custom_motd.sh"
-  end
-  
-  # Set AWS variables to be more appropriate for a development machine
-  ENV['TFB_AWS_EBS_TYPE'] = ENV.fetch('TFB_AWS_EBS_TYPE', 'standard')
-  ENV['TFB_AWS_EC2_TYPE'] = ENV.fetch('TFB_AWS_EC2_TYPE', 'm1.small')
-
-  provision_bootstrap(config, 'all')
-
-  provider_virtualbox(config, 'all')
-
-  provider_aws(config, 'integrated')
-  
-end
-
