--- deployment/vagrant-production/Vagrantfile
@@ -1,47 +1,0 @@
-# -*- mode: ruby -*-
-# vi: set ft=ruby :
-
-require_relative '../vagrant-common/checks'
-require_relative '../vagrant-common/core'
-provider = get_provider
-check_provider_needs(provider)
-
-Vagrant.configure("2") do |config|
-
-  server_ip = ENV.fetch('TFB_AWS_APP_IP', '172.16.0.16')
-  client_ip = ENV.fetch('TFB_AWS_LOAD_IP', '172.16.0.17')
-  databa_ip = ENV.fetch('TFB_AWS_DB_IP', '172.16.0.18')
-  
-  # Put the keys inside each box
-  Dir['keys/*'].each do |fname|
-    config.vm.provision :file do |file|
-      file.source = fname
-      file.destination = "~/.ssh/" + File.basename(fname)
-    end
-  end
-
-  config.vm.provision :file do |file|
-      file.source = "../vagrant-common/custom_motd.sh"
-      file.destination = "~/.custom_motd.sh"
-  end
-
-  config.vm.define "client" do |client|
-    provision_bootstrap(client, "client")
-    provider_aws(client, "loadgen", client_ip)
-    provider_virtualbox(client, "client", client_ip)
-  end
-
-  config.vm.define "db" do |db|
-    provision_bootstrap(db, "database")
-    provider_aws(db, "database", databa_ip)
-    provider_virtualbox(db, "database", databa_ip)
-  end
-
-  # Define the app server as the primary VM
-  config.vm.define "app", primary: true do |app|
-    provision_bootstrap(app, "server")
-    provider_aws(app, "appserver", server_ip)
-    provider_virtualbox(app, "server", server_ip)
-  end
-
-end
