--- deployment/vagrant-production/keys/client
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAqbwcjGjUhqqlTKt8CPiS7IyZr3AHV/ZY/lU5hqixDtX0rNUU
-LY0gSNNgDOLsk4refgZbqxUlUykcjhYv4Lbmkav7ru2cqkQcyGcXMvCluh4rZSCe
-1omCsu8RHCTVY0bR/A2/tZifmIEPMbqVxJYyYxUOeyTJfvO/SRGJDuoKNAEtpuYV
-JtvwOm6G6qHEAscrHD9aLrOVfkhUfhZ5Ikf5pehnxKI5J+i2uMGOsT4XS5BCVcMG
-ZDpy0EjNQCQN738fjbCuSWT6N+bURAKZpAayLOaJJzA0ckJwuBLN6+URnYeyT+JL
-/hAXSWC6YSdfGf7ZRzKlQ+P1keG+mnbU2rR8ZQIDAQABAoIBAB3AEgTuTun3uRdy
-K0BrSV1a24kriq35SOnzSCz5HmeqSsOMaYf+Z308R2aa07qcqZjPRnv7Ldqgpxek
-fenWRR2dVnT3wvD5FU5u+1r7YSD5LsP/7rdNzPHOA2K6dgh7nkyeoE6VHMwfOumN
-ebjGVriVNT+SWBr/YDTPdQV/MiChFyR2HgfT9wkDS4rtML1WpaQeNVHSyHNhjCdQ
-CHV8yLJGrPopXs2+DlD/QY3y9i6S6nFFLQCDgZYMW21LgtWIHdQ53xYnfR3qXrB6
-zqq5Qys0y283acRG/jwNlyu4o2K/Z95fqH9+WIn99j89UH89WUy8Ps4EM1yiik1g
-o61e+WECgYEA1jp1qJmwIz0jbV04Sf+IYspsGfdTBYxcQ3KcKfkUMOZ8t8pO7O96
-vY8JozZIyzWrh5yFZGqNBPKMAeBvFbaU3py+vBEHF7fpNsvo7TKXqx539wbThUPO
-uKrtw6+KQOa5vJfi8ASCvbzqD+hDQBseOEqiAGKJSZQy0WHMuImBMzkCgYEAytSw
-BHoMvK0wnJE045HpeM7l30skhHx/NaREUK/k0g/dc74C9Q6fwlDXCorhl+m+G3aX
-djZflOvYh60yajyjTyE5MBA9L2C8GldfjzbYYE02av3pQ+GGoN9wm2Z1FpKBU51m
-AwFme3AxkuObRhhl/7kGrmmzQCFMcijXup1ndo0CgYEAj+teOBJn3oSoVESYGD5v
-Zc8PqLIbOFL/eOB6QUhwZBm0F9T2CP8pukdp9hf1927q+YQRqVBbxMMdg/U/K5sp
-kBRFnxrhZPHM319F8HRYqEEj3vTbDWBsdGuAqEI54zgd97EiaJTfMJDFvkIPS/Xr
-621YQGHyUVgae8f0PNVD0yECgYAampMLw0ZwaPb7UkBxqO3Knrif9ScxCXNWAdnE
-C44jz1jomX+hzO0/e4L5KwYz+sCosYFkS/V1fwnDof/+1z1EX8M0fqzHzhReEB/E
-lRkEaJw+pFc4RBh4eVY3y9SHxWeZJAgBVOmwPyMBdUEkjIa0Z/xsXAja3IM+ZgUF
-wSeU8QKBgQDRv/nbrPc06JFYhGI8U+fE6Y2Qy1k5XQzZ+lAi7zHVmlS3qHpaltgM
-eU6JI34RByAofTppC4AmFI5y8Y/XYiD+yweDQzeI6jgnVyzBM/lNMdPkCtg/MoWt
-lqiAWjrXBHfTUNqtXyf8eQG8YB5Llv0Di3K8JbIS6LA2VCIje+aI0g==
------END RSA PRIVATE KEY-----
