--- deployment/vagrant-production/keys/client.pub
@@ -1,1 +1,0 @@
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCpvByMaNSGqqVMq3wI+JLsjJmvcAdX9lj+VTmGqLEO1fSs1RQtjSBI02AM4uyTit5+BlurFSVTKRyOFi/gtuaRq/uu7ZyqRBzIZxcy8KW6HitlIJ7WiYKy7xEcJNVjRtH8Db+1mJ+YgQ8xupXEljJjFQ57JMl+879JEYkO6go0AS2m5hUm2/A6bobqocQCxyscP1ous5V+SFR+FnkiR/ml6GfEojkn6La4wY6xPhdLkEJVwwZkOnLQSM1AJA3vfx+NsK5JZPo35tREApmkBrIs5oknMDRyQnC4Es3r5RGdh7JP4kv+EBdJYLphJ18Z/tlHMqVD4/WR4b6adtTatHxl hamiltont@Hamiltons-MacBook-Pro-101.local
