--- deployment/vagrant-production/keys/database
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAsA0K9e7N+A1CvdxVUfM1GWR9rM/Kyo703MRw9TZHdI15QgUU
-VwIGLoNyaGeQJMfbDY6k3QzC1y2DmnVkRgy55Q8LLXN7bfHcgme8Yie1fsK9YNrN
-nzfEbzwsO+8t9tR7IJufWQ36TDl0pskBOuArABFIh22ZMD4iKc1SWPqlZYHWdA4u
-LWlQ5ZmqOhUCCdtYyQoKq/IF2G0ahiYhpEmdf+zmTvHc7BZxz0wdHCPbAc3JMMBQ
-XcjgHEPb2qVqMzx5V0X5WH7w42Q0rYkr69efJJmjj86kL5ZIMrHhNPUjGdRyw6IO
-4WK8osBoOZnCfqr668F8Nso7heOM/f8EeCkJLwIDAQABAoIBAFDurgtBLQ7LidMe
-Z1q1I8P4no3DOMbF0WtsI0GXImJCZ2wiSad2FsFrTD0S3wEJn3osCT5Vj3y8P7PS
-I7bQQuU6dTj/zc5NUcHhTokRGC6d99F+xOhdCsDYOL2Uxtyh8bib0jVuZ5KDF01V
-OmuGgj4Vs/IOQE4dQjdVXHXrQ/3VXx0xsKgVmepNUBeTd1Vvxu6LPxPLcjj3ViC/
-fADuv23sL6kqSCSICDcCYxM2jmEMN3XTHVqBubm/V7x9QaS13EyoehzqzM0R4z7f
-hFtsgrtZJ+NautzgvVq4wcBFATzQXVl08vVN/gJmf4gq92GGGzo+HU3q0C7A+bkC
-0jdJ3YECgYEA3GVcGpmzbXEpVGgkaTcx3DwSGoiJ7tbuTS84F4I5p9uOILX2JDyu
-tIicfVb0YFtnZ7wJ7ggZZUdz9gWV4TIpmZsjDDre/7TJq53z2dQpvA8DLfw1G+KN
-WjNH/o5f00hKgLgNKaApkSiENvvmXlKqLSUOigk84uOXLKXp6MdwD8UCgYEAzH3D
-OY+UJ7RGgl8HckXBCYxn+9ajuOOBHvqR5BAGcJJ2LSYErugYpGeh8W46v+vq/7Ar
-jIAcV6hqDmG62ibOErIr13QMn0V0dz6Zgf20bbCcAYCLRcFP8kA2HfxP9ntUs48q
-M+J0sG+6efzDB54W2xi8Qz1jwrtwmpVKcUFGMGMCgYEAgnwcQL95JLWXw5p43DZK
-ddq22TC3gUmNAvOO4RQ/sGrjxegVV7T5tncfN3IQ8y+KlDcXJe9zoavhMupxp0d8
-q63W3TnJp1jRAaNdqTzcjLPSrRoiE2VZTOsJl+ORRlUJjxXo7RkfIwr65FbVOEj2
-6g2DgyQfATG/4jZEu4XAigUCgYEAnb82ZqT36VkMO24LVH+y1ibV4FRHoE0AQmGR
-QZmLHjQdrd5yJmss90ZqitA5yeu4MN45+fyp4IUuq0GA3uCof4uqfl4rjoIJwiMs
-UDbihOuErTP4PCtP/NTD6oXFOaMSSCHcsIG0+pZ2B7yQxsORZhrCwCwPVR07VERu
-VMfB2XkCgYA1Fp/pThnPQAOjYUftj9LSTj6Zm+IYZKyv3oXiO6IGjc4MWvpRmyj+
-G/gXbCZhL40z0hGIHMdyeu3ddQ+pBSncI0ycccbRGvPbITw9EXs+uzFflaYR/ocf
-v0hOmP/FvhuKH6shk8QjMPYENt5M5dRDZMujqtIgD+p40z98YqIxow==
------END RSA PRIVATE KEY-----
