--- deployment/vagrant-production/keys/database.pub
@@ -1,1 +1,0 @@
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCwDQr17s34DUK93FVR8zUZZH2sz8rKjvTcxHD1Nkd0jXlCBRRXAgYug3JoZ5Akx9sNjqTdDMLXLYOadWRGDLnlDwstc3tt8dyCZ7xiJ7V+wr1g2s2fN8RvPCw77y321Hsgm59ZDfpMOXSmyQE64CsAEUiHbZkwPiIpzVJY+qVlgdZ0Di4taVDlmao6FQIJ21jJCgqr8gXYbRqGJiGkSZ1/7OZO8dzsFnHPTB0cI9sBzckwwFBdyOAcQ9vapWozPHlXRflYfvDjZDStiSvr158kmaOPzqQvlkgyseE09SMZ1HLDog7hYryiwGg5mcJ+qvrrwXw2yjuF44z9/wR4KQkv hamiltont@Hamiltons-MacBook-Pro-101.local
