+++ deployment/vagrant/.gitignore
