+++ deployment/vagrant/custom_motd.sh
